// swiper_lazy_images/tp1/src/app/__models/misc.enum.ts
// ! Garder les clés des énumérations identiques entre AppComponentsEnum et XitiTagEnum

export enum AppComponentsEnum {
    ACCUEIL = "accueil",
    CAMPAGNE_DON = "campagne-don",
    TERMES = "termes",
    CERCLE_RICHELIEU = "cercle-richelieu",
    GALERIE_MAZARIN = "galerie-mazarin",
    RESULTATS_RECHERCHE = "resultats-recherche",
    SALLE_OVALE = "salle-ovale",
    SALON_LOUIS_XV = "salon-louis-xv",
    BREVIAIRE_CHARLES_V = "breviaire-charles-v",
    HOFFENBACH = "hoffenbach",
}

export enum CoursivesEnum {
    COURSIVE_RDC = "Coursive RDC",
    COURSIVE_1 = "Coursive 1",
    COURSIVE_2 = "Coursive 2",
    COURSIVE_3 = "Coursive 3",
}
export enum XitiTagEnum {
    ACCUEIL = "Page d'accueil Borne des Donateurs",
    CAMPAGNE_DON = "Campagne don Borne Des Donateurs",
    CERCLE_RICHELIEU = "Cercle Richelieu Borne Des Donateurs",
    GALERIE_MAZARIN = "Galerie Mazarin Borne Des Donateurs",
    RESULTATS_RECHERCHE = "Résultats Recherche Borne Des Donateurs",
    SALLE_OVALE = "Salle Ovale Borne Des Donateurs",
    SALON_LOUIS_XV = "Salon Louis XV Borne Des Donateurs",
    TERMES = "Termes Borne Des Donateurs",
}