// swiper_lazy_images/tp1/src/app/__models/pictures.enum.ts
export enum MatchersImgEnum {
    GALERIE_MAZARIN = "Galerie_Mazarin.jpg",
    SALON_LOUIS_XV = "Salon_Louis_XV.jpg",
    BREVIAIRE_CHARLES_V = "Accueil_bréviaire.jpg",
    HOFFENBACH = "Accueil_et_carrousel_5_Offenbach.jpg",
}

export enum CercleRichelieuImgEnum {
    IMG_0 = "Richelieu_Carousel_0.jpg",
    IMG_1 = "Richelieu_Carousel_1.jpg",
    IMG_2 = "Richelieu_Carousel_2.jpg",
    IMG_3 = "Richelieu_Carousel_3.jpg",
    IMG_4 = "Richelieu_Carousel_4.jpg",
    IMG_5 = "Richelieu_Carousel_5.jpg",
    IMG_6 = "Richelieu_Carousel_6.jpg",
}

export enum GalerieMazarinImgEnum {
    IMG_0 = "Mazarin_Carousel_0.jpg",
    IMG_1 = "Mazarin_Carousel_1.jpg",
    IMG_2 = "Mazarin_Carousel_2.jpg",
    IMG_3 = "Mazarin_Carousel_3.jpg",
    IMG_4 = "Mazarin_Carousel_4.jpg",
    IMG_5 = "Mazarin_Carousel_5.jpg",
    IMG_6 = "Mazarin_Carousel_6.jpg",
}

export enum SalleOvaleImgEnum {
    IMG_0 = "Ovale_Carousel_0.jpg",
    IMG_1 = "Ovale_Carousel_1.jpg",
    IMG_2 = "Ovale_Carousel_2.jpg",
    IMG_3 = "Ovale_Carousel_3.jpg",
    IMG_4 = "Ovale_Carousel_4.jpg",
}

export enum SalonLouisXVImgEnum {
    IMG_0 = "LouisXV_Carousel_0.jpg",
    IMG_1 = "LouisXV_Carousel_1.jpg",
    IMG_2 = "LouisXV_Carousel_2.jpg",
    IMG_3 = "LouisXV_Carousel_3.jpg",
    IMG_4 = "LouisXV_Carousel_4.jpg",
    IMG_5 = "LouisXV_Carousel_5.jpg",
    IMG_6 = "LouisXV_Carousel_6.jpg",
}

export enum BreviareCharlesVImgEnum {
    IMG_0 = "Accueil_carrousel_1_Breviare_CH-X_f320r_DSF1318hdV2.jpg",
    IMG_1 = "Accueil_carrousel_5_Breviare_CH-X_f304v_DSF1312hd.jpg",
    IMG_2 = "Carrousel_bréviaire_3.jpg",
    IMG_3 = "Carrousel_bréviaire_4.jpg",
}

export enum BreviareCharlesVImgSmallEnum {
    IMG_0 = "Accueil_carrousel_1_Breviare_CH-X_f320r_DSF1318hdV2-small.jpg",
    IMG_1 = "Accueil_carrousel_5_Breviare_CH-X_f304v_DSF1312hd-small.jpg",
    IMG_2 = "Carrousel_bréviaire_3-small.jpg",
    IMG_3 = "Carrousel_bréviaire_4-small.jpg",
}

export enum HoffenbachImgEnum {
    IMG_0 = "carrousel_offenbach_1.jpg",
    IMG_1 = "carrousel_offenbach_2.jpg",
    IMG_2 = "carrousel_offenbach_3.jpg",
    IMG_3 = "carrousel_offenbach_4.jpg",
    IMG_4 = "Accueil_et_carrousel_5_Offenbach.jpg",
}